Cope with long environment variables that are wrapped over multiple lines
The TERMCAP variable, if present, can be very long (27 lines and
over 1500 chars in my case).

In these cases, the earlier lines end with a backslash, and the
following ones start with a one tab indent.

Filter for lines starting with "export" to properly find the
variable names only, as that's the only thing we're interested in
here.